requirements.txt
google-cloud-bigquery==3.25.0
pandas==2.2.2
pyarrow==16.1.0
rdflib==7.0.0
requests==2.32.3
